Spike API to make the difference between configurations used for resolution and those used for selection
This commit spikes what an API to make the difference between a configuration being resolved and a configuration
used as a container for dependencies may look like. It introduces a `ConfigurationRole` enum, which is used
to make the difference between two use cases:

- `FOR_RESOLUTION`: when you resolve this configuration, select the `debug` build type of dependencies
- `FOR_SELECTION`: when this project/component is included in a dependency graph, here’s which configuration to use for `debug` build type

Typically, when resolving the `compileDebug` configuration, we set the attributes `buildType: 'debug'` to indicate that we want
to resolve to configurations which have the `buildType: 'debug'` attribute. This will resolve to `_compileDebug` configuration.
However, the target project itself needs to be able to have a `compileDebug` configuration, with the same attributes, to determine
the dependencies it requires when compiling its debug variant. This causes 2 distinct configurations to have the same
set of attributes. With this new API, we would only include the configurations that have the `FOR_SELECTION` role.

For backwards compatibility, the default is `FOR_RESOLUTION`. However, we might want to introduce a smarter
flag, `FOR_BOTH`.

See: https://github.com/gradle/performance/issues/135