Add an integration test illustrating how to implement variant aware dependency management
This test is an entry point for a collection of tests that will highlight how to implement variant-aware dependency
managenent with configuration attributes. It tries to simulate what the Android plugin does, and the first useful
test shows that we can make the difference between the dependencies used to compile a project and the dependencies
used when we compile _against_ a project.

In other words, this test implements the equivalent of "exported" dependencies in the software model, or in the Jigsaw
world:

- `client` depends on `core`
- `core` declares a compile dependency onto `commons-lang`, which is used internally
- when we compile `client`, the `commons-lang` dependency is not found on classpath

It is possible to add specific dependencies for compiling the project (`compileFreeDebug` configuration) or
dependencies found when we compile _against_ the project (`_compileFreeDebug`)

See: https://github.com/gradle/performance/issues/135