Attempt to simulate more complex Android setup
This commit tries to simulate how Android wants to avoid to build the `aar` when it just requires
compile dependencies. It introduces a new configuration for resources, and adds an additional
attribute to configurations, for usage.

The compile configurations have the attribute `usage: 'compile'`, while the resources configuration
have the attribute `usage: 'resources'`. We then have 2 kind of artifacts: the `aar` is a complete
artifact which needs the resources, wheareas the `jar` only contains classes (this is done to
simulate compile-only, in the end we want to depend on a directory).

When building the `jar`, resources are not processed.
When building the `aar`, resources are processed.

See: https://github.com/gradle/performance/issues/135