Makes impossible to create a project dependency onto a bucket configuration
If a configuration role is `BUCKET` then it is not supposed to be used as a
project dependency. This commit introduces an integration test to make sure
it is the case.

If a user wants to depend on such a configuration, it is always possible
to copy it.

See: gradle/performance#135